move mentions logic out to useMentions, move BareChatInput to its own directory
